use NSInteger for start_pos and end_pos instead of int.
NSInteger can be either int or long. While it's usually safe to cast convert int to long, casting long to int
(xv_0 returns NSInteger) is not safe

+++ XVim/NSTextView+VimMotion.m
@@ -1789,8 +1789,8 @@ NSRange xv_current_block(NSString *string, NSUInteger index, NSUInteger count, B
 		}
 	}
 	
-    int start_pos = (int)idx;
-    int end_pos   = (int)idx;
+    NSInteger start_pos = idx;
+    NSInteger end_pos   = idx;
     
     while (count-- > 0)
     {
